requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
scikit-learn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pandas
joblib